Update imports in tests
