+++ indra/tests/test_reading_scripts.py
@@ -7,11 +7,11 @@ import zlib
 from os import path
 from nose.plugins.attrib import attr
 
-from indra.tools.reading import read_db as rdb
-from indra.tools.reading.read_files import read_files
-from indra.tools.reading.script_tools import make_statements
-from indra.tools.reading.readers import SparserReader
-from indra.tools.reading.readers import get_readers as get_all_readers
+from indra.tools.reading.db_reading import read_db as rdb
+from indra.tools.reading.pmid_reading import read_files
+from indra.tools.reading.util import make_statements
+from indra.tools.reading.db_reading.readers import SparserReader
+from indra.tools.reading.db_reading.readers import get_readers as get_all_readers
 
 from indra.db import formats
 from test_db import get_db as get_test_db
